Fixed Bug #{4077}
The problem turns out to be the friction term.  The denominator of friction.n
depends only on non-empty groups.  The numerator depends on both empty and
non-empty groups, and thus can (in theory) grow without limit as empty groups
are added to the scenario.  But the denominator is meant to normalize the
result to the range -100 to +100, and clearly it can't do that if the 
numerator can grow without limit.

The solution is clearly to limit the numerator to non-empty groups.  The
numerator is

   sum(g in n, enemy.ng * force.ng)

We need to make that

   sum(g in n where own_force.ng > 0, enemy.ng*force.ng)

With this change, adding the empty group to the neighborhood doesn't change 
the security levels.