Fixed Bug #{4082}
Done!

From an operational viewpoint, nothing has changed. The logs appear in the
application just as they did before. Saving scenarios is identical as it was
before.  The only difference is that when running in batch mode you can specify
where the root of the working directory is by using the -scratch option.

The Athena PBS application makes use of this option.

Note: Athena appends a random 8 digit number as a directory to the root of the 
working directory to guarantee uniqueness.

Suggested tests:
Bring Athena up and at the command line type:
> super workdir join

Verify that the working directory returned is some directory under your home
directory if you're on Windows. On Linux, I think the directory is /tmp.

Launch athena from the command line and specify the -scratch option:

athena -scratch <some path you can write to>

Issue the same command as above and verify that the working directory is what
it should be.

Try using a non-existent directory as the argument to -scratch. 
Verify that an error.log is written. If you are running non-batch or in batch
mode on Linux, you should see a message written to standard out as well.