Fixed Bug #{4053}
Added clarifying material to the glossary entries for "support" and "vertical
relationship".  I didn't modify any of the Gofer pages because quite a few
of them reference either support or vertical relationships, and I'd have
to update all of them.  Instead, if the user wants to know what "support"
really means he'll click on it, and see the warning note.