Fixed Bug #{3881}
Fixed.  Whether the actor is an income or a budget actor is now shown in the 
overview table on my://app/actors; and on my://app/actor/{a}, it shows the
table of income by sources for income actors and the budget (with an
explanation of what that means) for budget actors.