Fixed Bug #{3873}
Added these columns to the hist_econ table:

subsisters -- the number of people engaged in subsistence agriculture
csf        -- the consumer security factor
rem        -- the amount of remittances 

To test: 
Run Athena for a while with a scenario. It would be best to have a remittance 
change rate applied in the econ model.

Point the detail browser at my://rdb and inspect the hist_econ table for 
the new columns.