Fixed Bug #{4094}
Fixed almost all of the problems listed in the bug; in addition, updated the
release notes page for Athena 5.

The Plot tab problem was ignored, because the Plot tab is never used and
will be going away in Athena 6.  The ordering of the priority, once, and
on_lock parameters on the tactic order pages was also left alone.