Fixed Bug #{4108}
gofer::CIVGROUPS now ensures that non-empty groups are excluded from its 
return values.

Note that the user can include empty groups in the BY_VALUE rule's
raw_value; they might not be empty any more when the gofer is evaluated.
But if they are empty, they will be excluded when the gofer value is
evaluated.