Fixed Bug #{4118}
Fixed!

The application now deletes all history from a supplied axdb with prejudice.

This means that if a user picks an AXDB that has history by accident, then it
gets blown away. Of course, if the same number of test cases are run against
the same baseline scenario, the results and history are simply reconstructed.