Fixed Bug #{4116}
Fixed!

This is really a problem only with trying to request a run with more tests than
are present in the AXDB. Requesting more nodes than there are tests simply 
means using the same number of nodes as there are tests, it's not an error unless
you try to request more nodes than PBS can support.

The dialog now validates the number of tests requested and if it's more than is
possible, it is flagged as an error.

Retest.