Fixed Bug #{4135}
Done!

Geographic unemployment (GU) is now an input to the CGE and effects the size of the
economy and the unemployment rate.

Since GU is computed outside the CGE in a model that is yet to be implemented,
there is only one way to cause GU to change: by using the Update CGE Cell Value
order under the Orders > Economics menu.

To change GU, which can only be done after the scenario is locked, use In::GU
as cell name.

Remittances are now also an input to the SAM instead of being computed. If 
remittances are non-zero, it is up to the analyst to determine to which sectors
the remittances are allocated. Otherwise, remittances simply go out to the
world when the pop sector is balanced.

Suggested tests:

  * Input some value for remittances on the SAM tab and do not allocate any of
    it to other sectors and lock the scenario.
  * Verify that remittances simply flow back out to the world in the T-matrix.
  * Unlock and allocate some (or all) the remittances to some other sector 
    (BX.goods.pop is a good one) and lock the scenario.
  * Verify that some (or all) of the remittance remain in the economy and GDP
    is increased.
  * After the scenario is locked, use the Update CGE Cell Value order to change
    In::GU. Verify that the economy is affected by the change in the number of
    people working (GDP goes up/down, Unemployment goes up/down, etc...)
  * Note that you only see these changes on the "Constrained with Geo. Unemp
    and Sec. Factors" view of the economy.