Fixed Bug #{4137}
Done.

This means that the GDP and the unemployment rate on the "M" page will no 
longer match those from the Base SAM. However, the "Calibrated" view still 
will. Since the calibrated view is now no longer a duplicate view 
(ie. no longer matches the "M" page) that view is renamed "Calibrated Values 
from Base SAM" and when that view is active on the Econ/Overview tab the
prose at the top reflects that. It no longer mentions that the view is for 
debugging only, since it is now a view that may be useful to an analyst.

Tests:
 * Using the Update CGE Cell Value order in the Orders > Economics menu change
   the value of In::GU to different values after locking the scenario.
 * Verify that the "Constrained with Geo. Unemp." view changes.
 * Verify that the "Calibrated Values from Base SAM" view does not.
 * Note the prose for each view and verify that it makes sense.
 * Note: the "Constrained with Geo. Unemp. and Sec. Factors" view will also 
   change when In::GU is changed.