Fixed Bug #{4146}
Done!

The Tcl [info] command provides some information about each stack frame.
The bgerror proc in the Athena application has been updated to use it as an
aid to debugging bgerrors.

Testing this is a bit weird since a bgerror must happen to see the new info.
Fortunately, you can use the 'super' command at the CLI to just call the 
bgerror proc directly:

> super bgerror foo

('foo' can be any string)

You should see the log pop up and a nicely formatted bgerror context with 
stack level and caller should appear in orange in addition to the bgerror that
is already produced.