Fixed Bug #{4147}
Done.

The helps and the UI have been updated to display the new units.

Black market feedstock units are now blackFeedBKT, which is just a notional
set of black market feedstock (eg. for opium it would be poppies) from which
a blackBKT is produced.

Verify that the helps for the Econ tab are updated and verify that the UI
has blackBKT now.