Fixed Bug #{4157}
I was making an unwarranted assumption that tactics executed "on lock"
never fail to obligate all of the resources they need.  That's true for 
cash, but not true for personnel.  I've revised the DEPLOY tactic accordingly,
and fixed a related issue in block(sim), which had an explicit assertion 
that tactics always obligated successfully on lock.  As a result, the
block code is a little simpler.