Fixed Bug #{4188}
Done!

The SPEND tactic now has the same modes as the DEPOSIT tactic:

ALL (default): Spend all cash-on-hand
EXACT: Spend exactly this amount of cash-on-hand
UPTO: Spend up to this amount of cash-on-hand
PERCENT: Spend this percentage of cash-on-hand
EXCESS: Speand any amount of cash-on-hand in excess of this amount

Tests:
* Create a strategy block and a SPEND tactic within that block
* Verify that by default the ALL mode is used
* Create different SPEND tactics using different modes
* Verify that for each mode the proper amount of cash-on-hand is spent

Note: only the EXACT mode can result in a tactic execution failure, and only 
if the tactic is executing on a tick as opposed to on-lock.

TBD: Update the docs.