Fixed Bug #{4197}
Added a "validate" method to beanclass(n).  This means that every beanclass
object now has a validate method that will efficiently validate an ID of
that beanclass, throwing an appropriate error if it's invalid.

Updated all beanclass-related orders in Athena to use

   prepare id -type $beanclass

to validate bean IDs.  This included updating their test scripts.