Fixed Bug #{4187}
Had to relax the validation in the TACTIC:CURSE order so that Athena state
changes between a cut and paste of a CURSE tactic doesn't fail.

Because of this, the sanity check in the CURSE tactic has been beefed up and
the test suite updated.

Suggested tests:
* Use the GUI to create some valid CURSE tactics
* Use the "send" command at the CLI to send bogus data for the CURSE
* Verify that it is accepted, but that the CURSE is marked invalid.
* Verify that you can attempt to lock the scenario but it will fail with an appropriate sanity check error message.

Note: consult the help for the TACTIC:CURSE order to assist with constructing "send" command