Fixed Bug #{4203}
Done!

The service module now takes into account the urbanization of the neighborhood
a civilian group is in when computing the amount of funding received.

This is going to effect the FUNDENI tactic once it is converted to the Mark II
strategy architecture. 

It will also be more readily testable when that tactic is brought up to date.

See Bug 4193 for test suggestions.

The automated test suite has also been updated.