Fixed Bug #{4193}
Done!

The FUNDENI tactic is now available as part of the Mark II strategy. The tactic
has the same modes as the DEPOSIT tactic (and others). It also allows the user
to specify a cap on spending as a function of percentage of saturation level of
service (SLOS). This percentage defaults to 100%. Setting it to something less
will always cause ENI funding to be no more than the amount of money it would 
take to reach that level of funding assuming that the actor that owns the 
tactic is the only one funding ENI.

Tests:

* Create a strategy block and add FUNDENI tactics to it.
* Use the different modes to verify that the amount of money expended makes sense
* Use different SLOS percentages to cap funding.

Note that the funding of ENI is no longer based solely on population. The 
urbanization of the neighborhood in which a funded civilian group resides is
taken into consideration. This is consistent with the computation of funding
level based on a percentage of SLOS.

The automated test suite is updated.