Fixed Bug #{4215}
The problem was that the idCounter wasn't being reset when undoing the
creation of bean.

All beans created via orders were created by mutators that added them to
slots.  There was an inconsistency here: [$bean addbean_] was given the
new bean, created by its caller, but was responsible for uncreating it on
undo.  To avoid mismatches, then, I did the following:

* [$bean addbean_] takes the class of the bean to add, and creates the bean
  itself.
* It also now takes an optional variable name; the variable will receive the
  name of the newly added bean.
* On undo, it calls a new command, [bean uncreate], which destroys the bean
  and decrements the idCounter.
* [bean uncreate] verifies that the bean being destroyed really is the 
  newest bean; thus, if our undo scripts aren't in the right order, we'll
  soon know.
* Updated each of the uses of addbean_ to use the new calling sequence.

The paste_ code is currently broken; but we weren't going to use it in 
its current form anyway.