Fixed Bug #{4217}
At this point, the DEPLOY tactic contains all of the required changes
except for the three new pmodes; in addition, the help pages have not
yet been written.  I'm committing what I've got to enable Dave Hanks
to proceeded.  The new pmodes and help will appear in a subsequent build.