Fixed Bug #{4243}
This commit represents the first cut. This package will allow the user to:

* Specify a production capacity fraction by neighborhood.
* Specify how actors are alloted shares of the manufacturing infrastrucure.
* Define initial repair levels of the infrastructure
* Let the SYSTEM agent own any or all of the infrastrucure.

With this information Athena will:

* Determine the number and allocation of plants based on:
  ** Individual plant production capacity.
  ** The initial demand for goods in the goods sector from the calibrated CGE.
  ** The production capacity factor in each neighborhood.
  ** The initial repair level of plants owned by agents.

The newly introduced "plant" module will then compute production capacity based
on the number of plants in the simulation and the repair level.

The econ model will pull in that capacity from the plant module at each tick.

This commit contains only the basic functionality, it does *not* contain:
* Degradation of plants
* The MAINTAIN tactic
* The BUILD tactic.

If the user chooses to do so, any neighborhood that has production capacity
but no actors are defined to receive shares of that capacity, the SYSTEM agent
will be allocated all the plants needed to meet that capacity. This means that
by simply doing nothing, the SYSTEM will be allocated all the plants in every
neighborhood.

In the future, plants owned by the SYSTEM will not need to be maintained by any
tactic. They will not degrade over time. This means they are effectively "auto-
maintained."

Suggested tests:
* Access the allocation of shares on the Neighborhoods > Manufacturing Plants
tab and assign shares of production capacity to various actors in various 
neighborhoods.
* Verify that in the order dialog, it is not possible to create a new 
allocation of shares in a neighborhood for an actor that is already allocated
shares in that neighborhood.
* Use the Neighborhood browser to change the production capacity factor.
* After locking the scenario, access the detail browser for both actors and
neighborhoods to see exactly how manufacturing plants have been allocated.
* Play with the econ.bktsPerYear.goods parameter and change the output of a
single plant to cause more or less plants to be created.
* Use different initial repair levels.
* Verify that, no matter what, after the scenario is locked the sum of the 
number of plants multiplied by the repair level of the plants should be equal
to the quantity of goods baskets demanded in the calibrated view of the
economy on the econ tab:

      QD.goods  = sum over all n,a (plants_na * rho_na) 

* Inspect detail browser pages for PLANTS:* orders and manufacturing plant
objects.

There's probably a lot that could be broken, so break it!

Still TBD: automated test suite tests.