Fixed Bug #{4270}
Fixed.

The model adjusts the number of plants an actor initially starts with based
on repair level in an attempt to fill exactly the capacity in the goods sector
that the econ model calibrates.

In the case of an initial repair level of 0.0, no plants are producing goods so
there's no way to adjust the number to anything meaningful. Thus, the adjustment
does not take place in this case and the economy starts in a degraded state due
to the non-functioning plants, which is what you would expect.

Retest.