Fixed Bug #{4287}
The nlist attribute is optional in the sense that the order processor
does not reject a TACTIC:MAINTAIN order when it is missing.
It does not mean that the tactic itself is valid without it.

All other tactics that have gofers behave this way as well, which may be wrong,
but it is this way to support the cut and paste of tactics when the scenario
changes between the cut and the paste in such a way that a tactic becomes 
invalid.

The helps have been updated so that they match the order dialog.