Fixed Bug #{4289}
This is now fixed. 

During simulation bootstrap the econ module was starting up after the plant
module, which means that the default value for calibrated QD.goods was 
being extracted for use by the plant module instead of the computed value.

Retest.

There's a couple other things being included in this commit:
* When the econ model is disabled, so is the infrastructure model.
* The infrastructure detail browser page now includes a link to the SYSTEM agent
  instead of just plain text.

Disable the econ model and verify that the detail browser pages for the 
infrastructure informs the user that the infrastructure model is disabled.
Also verify that the detail browser pages for individual actors and 
neighborhoods make mention of this in their infrastructure sections.