Fixed Bug #{4286}
I modified executive.tcl and gofer_number.tcl to change the order so instead
of assigned(n,g,activity), it is assigned(g,activity,n).
I modified gofer_number.tcl so the ASSIGNED gofer::NUMBER rule uses the activity
check to validate the activity based on it's group type.

I modified activity.tcl so that it's ::activity asched names typemethod also
takes an optional group {g} argument. If not given, it returns all assignable
names. If g is given it returns the assignable names for that gtype.

I modified the test 010-activity.tcl so that it has an added test to test the
optional group argument.

I modified the test 030-function.test and 010-gofer_number.test to take the 
arguments in the new order (g,activity,n) instead of (n,g,activity).

This is done.

To Test, Follow the test instructions in bug 4276. Verify the format is now 
(g,activity,n).