Fixed Bug #{4295}
Done!

The SYSTEM agent can add ATTRIT tactics to it's strategy.

The ATTRIT tactic has two modes:
* NBHOOD (default)
* GROUP

NBHOOD mode causes magic attrition to units of all groups in a neighborhood.
GROUP mode causes magic attrition to units of the specified group.

The tactic can also be created with responsible groups specified
just as it was in the old orders.

All references to the old orders have been removed from the GUI and
docs.

The docs have been updated to contain the new tactic and the test suite
is updated to remove obsolete tests and add new ones.

Test all aspects of the ATTRIT tactic. Note that since the strategy is 
set up during PREP, it's possible to specify a tactic in GROUP mode
for a group that does not have units in the chosen neighborhood.

In this case, the tactic does not fail, it just does nothing. In other
words, this tactic works on a best efforts basis.