Fixed Bug #{4300}
This is fixed.

If you lock the scenario, find a group that you know is NOT in the first nbhood, 
N00 or NB0, etc.  

From the console do "security yourgrp nbhoodItIsIn". That is the correct value. 
Now do "security yourgrp firstNB". It should be 0.
Now do "security yourgrp". It should be the correct value of the first command above.