Fixed Bug #{4290}
Revised the mechanism for controlling the "Add Tactic" pulldown menu
on the strategybrowser "tactic list" component.  The menu items are now
controlled using a predicate that retrieves the valid tactic types for
the given agent from the [agent] module.  That module builds the list
given the agent type, but is then free to customize it, e.g., to remove
the MAINTAIN tactic for actors with "auto_maintain" set.