Fixed Bug #{4254}
Fixed.  We now distinguish between tactic failures during planning, i.e.,
during obligation, and tactic failures during execution, i.e., after the
actors' decisions can no longer be altered.  The former should have
a red failure icon in the eresource type, and the latter should have a
a purple failure icon.  In addition, the failure message in the latter
case should begin with the words "Failed during execution:".

The BROADCAST tactic now works this way: CAP access is checked at 
assessment time, at the end of strategy execution, as it must be, and 
updates the tactic's status appropriately if access is not available.