Fixed Bug #{4315}
Done!

The DAMAGE tactic is only available to the SYSTEM agent.

It can be used to both increase or decrease the average repair level of
infrastructure.

Since DAMAGE tactics are specified during PREP, it's possible that one could
be set up for an actor/neighborhood pair that does not correspond to any
infrastructure ownership at the time it executes.

In this case, the tactic executes trivially and a sigevent is logged with
an appropriate message.

The online helps and automated test suite have been updated.