Fixed Bug #{4324}
Done.

When allocating shares of infrastructure, you can no longer specify non-local neighborhoods.

Also, when creating or editing neighborhoods, PCF is required as an input to the order dialog only when Local is "Yes".

It is possible to change a neighborhood from local to non-local after allocating infrastructure to it.

In this case, a sanity check failure should occur on lock laying out the problem.

Tests:
* Create some non-local and local neighborhoods. Verify that PCF is required only when Local is Yes.
* Update some non-local and local neighborhoods. Verify the same.
* Verify that you cannot allocate shares of infrastructure to non-local neighborhoods.
* Verify that a sanity check failure occurs if a neighborhood that has infrastructure allocated to it is changed to non-local.
* Verify that GOODS production infrastructure is allocated appropriately (ie. its like non-local neighborhoods do not exist for the purpose of allocation)
* Verify that the infrastructure section in the detail browser pages for non-local neighborhoods makes sense.
* Verify that "N/A" appears in the detail and SQL browser under PCF for non-local neighborhoods.

The test suite has been updated.