Fixed Bug #{4342}
Disaggregation of unemployment is done by the demographics model.

Workers that are "too far" from jobs will not travel to neighborhoods that have them.

There is a parmdb(5) parameter that controls this maximum commute called demog.maxcommute.
By default it is set to "FAR".  Setting this parameter to "REMOTE" would make all workers available to all neighborhoods to fill jobs.

The input number of jobs to the disaggregation model is the capacity limited number of jobs from the CGE's "M" page.
This number of jobs is limited by both the available workforce and the GOODS manufacturing infrastructure capacity in the infrastructure model.

This number of jobs is divided up by the ratio of each neighborhoods GOODS capacity to the playbox total capacity.
Then jobs are allocated to neighborhoods filling lower priority (HERE) first, then moving to NEAR and then FAR and REMOTE if the demog.maxcommute parm indicates it should.

Once jobs are disaggregated, if the unemployment computed by the demog model is larger than the unemployment computed on the "M" page, the difference is geographic unemployment.

The geographic unemployment is an input to the "S" page where people that are unwilling (or unable) to travel to work further impacts the economy.

Some thigs to keep in mind when testing:
* The unemployment on the S page ("Constrained with Geo. Unemp. and Sec. Factors" on the GUI) should always be greater than or equal to that on the M page ("Constrained by Labor and Goods Capacities" on the GUI).
* The GDP on the S page should always be less than or equal to that on the M page.
* Setting demog.maxprox to REMOTE makes geographic unemployment always zero.

The test suite has been updated. The documentation still needs to be done.