Fixed Bug #{4358}
With the changes committed under bug 4357 the place that information for GOODS production infrastructure exists on an actor by actor basis has changed significantly.
Because of this, it didn't make sense to implement what was suggested. Rather, the link "Infrastrucure Ownership" is changed to "GOODS Plant Ownership" so as to accurately describe what's there.

In the actor specific section there is some verbiage that discusses GOODS production ownership in high level terms with a link to the actor specific page.

See if this answers the mail.