Fixed Bug #{4360}
Local neighborhoods that have no consumers now show up in the Infrastructure table in the Detail Browser.

They are shown as having 0% of the plants.

Upon lock, however, there should be no records shown for those neighborhoods that have 0% of plants allocated to them.

The headers have been fixed. 

Note, however, that the demographic tables, which actually contain the number of consumers, do not get populated until scenario lock, so what is shown in the detail browser is an interpretation that subsisters are not consumers.

If ever there becomes a reason to introduce a new class of population that is also not a consumer (not sure what this would be, exactly) then the database views the detail browser uses may need to be updated.

Retest.