Fixed Bug #{4235}
Done.

There are two modules that need to use the "money" type to validate parmdb(5) parameters that may have been input by the user in the money format.

To test:
Set these parmdb parameters to money formatted values:
activity.FRC.*.cost
activity.ORG.*.cost
service.ENI.saturationCost.*

And then execute the ASSIGN and FUNDENI tactics so that those parameters are accessed.
(ie. the force activity that has a cost is the activity in the ASSIGN tactic)

There should be no bgerrors.