Fixed Bug #{4416}
Done.

For now the only supported file format is GeoTIFF with GEOGRAPHIC model type.

This is a simple rectangular projection with tiepoint and pixel scaling.

Included with this commit is a sample GeoTIFF of Afghanistan to use in testing.

Tests:
Bring up Athena and import Afghanistan.tif from the maps directory.
Moving the mouse around the map should produce a display of MGRS and lat, long coordinate pairs in the map viewer's toolbar.
Use Google Maps or some other independent source of your choice to verify that the locations under the mouse are reasonably correct (I believe that this map is slightly off in its geo-reference metadata.)

Undo the import, verify that Athena goes back to using the homegrown way of displaying map coordinates.

Redo the import, and verify that the Afghanistan map is displayed with correct coordinates again.