Fixed Bug #{4465}
This commit contains the initial capability. Most of the work was done on an
experimental branch so a record of the incremental commits can be found there.

The WMS client can be launched from the File menu. The help pages contain a 
section on how to operate the WMS client so please consult there for a 
description of how to use the client.