Fixed Bug #{4476}
Since testing the wmsclient(n) requires an http server and we do not want to depend on the presence of a network, the automated tests use an httpagent(n) simulator.

No network connection is required, but it behaves like it is a full up http server.

Some XML test files that simulate the return of data from the server are also included.

No I&T is necessary as these are run automatically at build time.