Fixed Bug #{4477}
Done!

The wmsclient(n) object will use the default parameters if no explicit request has yet been made (ie. upon first connection to the server).

Otherwise, if no parameters are passed, the parameters used in the last request are used.

This change should be invisible to testing, it simply makes the wmsclient(n) API more robust.

Use the WMS window in the Athena application to request a few maps to make sure nothing is broken.