Fixed Bug #{4469}
There were two issues here.

1) bsys.tcl has a [bsys clear] command that clears all defined belief systems
and topics, and defines the "Neutral" belief system.  BUT it was using 
the [mutate add] mutator to create "Neutral", which sends out a notifier event.
It shouldn't; the [bsys clear] command is not a mutator, and is used only as
part of some larger event.

When unlocking the scenario, the belief systems are momentarily cleared and then
restored; and on the clear, the browser tried to update itself and got confused.

2) bsysbrowser.tcl combines the topic view data with the belief view data for
its topic/belief list.  If there is no belief, it puts in empty fields.  But
it didn't put in all of them.