Fixed Bug #{4482}
Done!

Now when the econ model is disabled the CONSUMP rule set will fire with inputs coming from the parmdb.
When the economic model is enabled, the inputs will come from the demographics model.

To test:
Create a scenario and disable the econ model.
Lock the scenario and use the detail browser to verify that the CONSUMP rule set fires.
To change the magnitudes of the effects, vary the dam.CONSUMP.expectf.* and dam.CONSUMP.povfrac.* parmdb parameters for different urbanizations.