Fixed Bug #{4483}
Done!

The new parameter is called demog.playboxUR and has a valid range of 0.0 to 100.0.

To test:
Create a scenario and disable the econ model.
Leave playbox UR at the default of 0.0 and lock the scenario. You should see neighborhood and group unemployment rates in line with the turbulence (nominally 4.0%).
Change the playbox UR (no need to return to prep) and verify that the neighborhood and group unemployment rates are in line.
Verify that the UNEMP ruleset fires with the correct numbers (unemployment attitude factor or UAF) is what's used.