Fixed Bug #{4490}
Done!

There is now an Economics/Control tag that contains a big, unmistakeable switch that allows a user to turn the econ model on and off and will in scenario prep.

When the switch is toggled, some text is displayed informing the user of what it means for the model to be on or off.

There is also a new detail browser page that gives a user step-by-step instructions on how to set up a very simple SAM that requires a minimum of data.

There is a link from the Control tab to this page when the econ model is on.

When the econ model is off, there is some text about what model parameters can be used to affect the UNEMP and CONSUMP rule sets.

The test suite has been updated. The parmdb parameter econ.disable is gone. The econ model state is stored within the econ model itself.

To test: 
* play around with the switch and verify that the econ model can indeed be turned on and off.
* read the text and test the links in the Control tab.