Fixed Bug #{4502}
Done!

WMS servers that are successfully connected are automatically remembered when Athena is restarted.
The "Server:" bar in the WMS window is now an editable combobox that allows the user to type in a URL or use the dropdown to retrieve URLs for servers that were successfully connected to in the past.

By default, the CUBEWerx server is the only URL in the list of URLs so it should appear in the dropdown.

Tests:

Bring up the WMS client window and verify that the CUBEWerx server is still the default server and that a default map is automatically retrieved.
There is another WMS server at this URL:
http://107.20.228.18/ArcGIS/services/Wetlands/MapServer/WMSServer

Type (or cut and paste) this URL into the "Server:" bar and press Enter.
Verify that the WMS window retrieves a default map and that you can interact with the map with bounding box and layers (there's not many).

After the server has been successfully connected, verify that the drop down contains the new URL.

Shut down and restart Athena. Verify that the new server is available in the dropdown.

Try invalid URLs or URLs that are valid, but not WMS servers and verify that appropriate error messages are generated and that invalid servers are not remembered.