Fixed Bug #{4510}
Done!

Use the "Export Scenario As..." menu item in the File menu to export a loaded .adb
Verify that the exported script contains a MAP:IMPORT:DATA order.

Ingest the generated order script into Athena and verify that the map and projection (if there was one) is correctly represented in Athena.