Fixed Bug #{4511}
Done!

This commit allows a user to select a set of files to be used for ingestion into the application.

This may be changing in subsequent commits since things are in flux.

Included is 25 TIGR messages that can be used for ingestion.

Time stamps are converted to week strings, locations are converted to neighborhoods.

The tigr module has a call to return any errors encountered during the message parsing step.
If an error is encountered, it is stored and parsing continues. 

The list of offending files and their errors (if there are any) can be retrieved through the tigr API.