Fixed Bug #{4513}
Fixed.

Turns out that it doesn't make any sense to try to compute ALOC, ELOC and Poverty % from a user supplied expectf. It's not possible to come up with numbers that make sense.

Ultimately, this tab will become invisible to the user when the economic model is disabled.

For now, the browser has two views: one for when the economic model is enabled, and one for when it is disabled.

When disabled, the columns that do not make sense are invisible to the user.

Retest.