Fixed Bug #{4492}
This particular problem is fixed. However...

The PCF isn't as useful as it once was. When the econ model is enabled. Production capacity is determined by the number of plants and thier repair level.
It seems that a different view is needed.

Retest and make sure the bgerror does not occur. 

See bug 4518 for more.