Fixed Bug #{4540}
Fixed.

This is now a sanity check failure should Athena detect that there is a conflict.
It's not an error to enter the same role in different role types because it's possible that during prep things are in flux as the analyst may be changing roles around and being presented with a bunch of errors that are going to go away eventually would be annoying.
So it is an error only if there are conflicts at lock or when a sanity check is explicitly asked for.


To test:
Rerun the previous test as described and try other cases.
In all cases where there exists a conflict, the sanity check should fail and report to the user what the problem is.
Precedence of role type is given to injects that appear first in the CURSE.

Try using the "Check" button on the CURSE browser to cause and clear sanity check failures.

The test suite has been updated.