Fixed Bug #{4526}
Done.

There is a history table called hist_service_g that now contains a history of ENI funding.

Whether the history is saved is controlled by a new parmdb parameter called hist.service which is boolean and defaults to true.

To test:
Bring up a scenario and advance time a few weeks.
Inspect the hist_service_g table and verify that it is populated:
 > select * from hist_service_g
Return to prep and set hist.service to "no".
Advance time again and verify that hist_service_g has no entries.