Fixed Bug #{4567}
The problem here is that the scenario lock/unlock button got set to locked
while in the WIZARD state, which is wrong.  I reviewed the appwin/SimState
routine and fixed up the logic to take WIZARD into account.

In particular, when in the WIZARD state the lock/unlock icon should show
an unlocked padlock, and the icon should be disabled (grayed out).