Fixed Bug #{4584}
The new order has been implemented and is called NBHOOD:CREATE:RAW.

This order is meant to be used by other software (such as the neighborhood ingestion wizard) but can be sent from the command line like any other order.

See the documentation of the order and try creating some polygons using raw lat/long data and a geo-reference map.