Fixed Bug #{4582}
The intel ingestion wizard now keeps track of how many completed ingestions
it has done for each scenario, and uses that to number events.  The events
of the first ingestion are number 1-1, 1-2, 1-3, ...; the events of the
second ingestions are numbered 2-1, 2-2, 2-3, ...; and so on.  Undoing an
ingesting doesn't change this numbering; if ingestion 1 is completed, undone,
a second is begun, its events will be number 2-1, 2-2, etc.