Fixed Bug #{4585}
Fixed a couple of bugs:

The outline of pre-existing polygons is not showing up because the fill color is the same color.
When selecting all children of a parent, the parent should become deselected (checkbutton unchecked) but only if the parent does not already exist in the scenario.

Tests:
Select some polygons and send them to the scenario.
Relaunch the wizard and verify that you can see the outline of the existing polygons.
Select all children of a parent and verify that if the parent does not already exist in the scenario, it becomes deselected.

NOTE: Because parent polygons can be selected and then, later, completely covered by children (and vice-versa), it's possible that refpoints become obscured. In this event, Athena proper must be used to fix the problems whether that means deleting unnecessary polygons or moving ref points.