Fixed Bug #{4590}
This is now fixed.

If there are no neighborhoods available because the map selected is outside the bounding box of all neighborhoods a message box is displayed stating so.
It also will show the bounding box of available neighborhoods as a guide.

The other part of this bug is also fixed, but with caveats:

It's possible to zoom way in (try the city of Kabul) and still get a neighborhood outside the "playbox" (in this case Pakistan should show up).

The reason for this is how the geometry of the playbox and neighborhoods are checked against each other.
A neighborhood is said to be "in view" if any of the following are true:

* The playbox is entirely contained within a neighborhood bounding box.
* A neighborhood bounding box is entirely contained within the playbox.
* At least two coordinates of a neighborhood bounding box are inside the playbox
* At least one vertex of neighborhood polygon is contained within the playbox.

The reason for the first three checks is that they are fast and good for a "normal" case where the playbox contains many neighborhood polygons.
But when the "playbox" is small compared to the neighborhoods it's possible for it to be completely contained within the bounding box of a neighborhood that is much larger than it but whose vertices are all outside it

The bottom line is that if you are going to be playing district or provincial sized neighborhoods then the size of your playbox should reflect that.

If it turns out that this isn't good enough then I may just use the more expensive "point inside polygon" check for the case when the playbox is small and outside a neighborhood's vertices.