Fixed Bug #{4592}
This turned out to be a significantly more complex issue than I originally thought.

The problem is that when a parent has some of it's children selected, and those children have children where some are selected, it's tricky to determine a viable ref point for the top parent.

Essentially, what has to happen is that the tree of neighborhoods has to be traversed to the leaves starting with the parent in question and then those leaves accumulated and inspected for a viable child to use for determining a reference point.

This bug occurred because the code was not doing a deep search of the tree for viable leaf nodes and, thus, no ref point was found.

The problem should be fixed now that the code is doing this deep traversal looking for viable children to use for ref points.

Retest and test some more!